Create file structure, mainly the 1st recipe page
Add index file, images folder, recipes folder, and README file to
the directory. The recipes folder contains the coconut-curry-tofu.html
file and the images folder contains tofu.jpg.

index.html contains a link to the coconut-curry-tofu.html file, which
includes a picture, tofu.jpg, and contains the ingredients list and
steps for the recipe.