Add black bean soup recipe page with content
Add black-bean-soup.jpg to images folder and black-bean-soup.html to
recipes folder.

The soup recipe page uses the same format as the tofu page, containing
both the ingredients, steps, and an image of the dish. The links to
these recipes have also been put into an unordered list within
index.html, so the aren't all on one line.