Add chocolate protein bites recipe page w/ content + Modify descriptions
Add protein-bites.webp to images folder and chocolate-protein-bites.html
to recipes folder.

The protein bites recipe page uses the same format as the other recipe
pages. The descriptions for all recipe pages were also changed to
include the prep and cook time. Finally, the link to the protein recipe
was added to index.html.